# packages/core/api.yaml
openapi: 3.0.0
paths:
...
  /v1/auth/apple:
    get:
      operationId: authAppleFindStrategiesV1
      parameters:
        - name: admin
          required: true
          in: query
          schema:
            type: boolean
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicAuthStrategyDto'
        '429':
          description: Rate limited
      tags: &ref_5
        - Auth
  /v1/auth/apple/{strategy}/login:
    post:
      operationId: authAppleLoginRedeemV1
      parameters:
        - name: strategy
          required: true
          in: path
          schema:
            type: string
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthAppleLoginRedeemDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSessionTokensDto'
        '429':
          description: Rate limited
      tags: *ref_5
  /v2/auth/apple/{strategy}/exchange:
    get:
      operationId: authAppleExchangeStartV2
      parameters:
        - name: strategy
          required: true
          in: path
          schema:
            type: string
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicAuthAppleExchangeDto'
        '429':
          description: Rate limited
      tags: *ref_5
    post:
      operationId: authAppleExchangeCallbackV2
      parameters:
        - name: strategy
          required: true
          in: path
          schema:
            type: string
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthAppleExchangeCallbackDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSessionTokensDto'
        '429':
          description: Rate limited
      tags: *ref_5
  /v1/auth/apple/{strategy}:
    get:
      operationId: authAppleSetupV1
      parameters:
        - name: strategy
          required: true
          in: path
          schema:
            type: string
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicAuthAppleExchangeDto'
        '429':
          description: Rate limited
      tags: *ref_5
    post:
      operationId: authAppleExchangeV1
      parameters:
        - name: strategy
          required: true
          in: path
          schema:
            type: string
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthAppleExchangeCallbackDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSessionTokensDto'
        '429':
          description: Rate limited
      tags: *ref_5
  /v1/auth/google:
    get:
      operationId: authGoogleFindStrategiesV1
      parameters:
        - name: admin
          required: true
          in: query
          schema:
            type: boolean
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicAuthStrategyDto'
        '429':
          description: Rate limited
      tags: &ref_6
        - Auth
  /v1/auth/google/{strategy}/login:
    post:
      operationId: authGoogleLoginRedeemV1
      parameters:
        - name: strategy
          required: true
          in: path
          schema:
            type: string
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthGoogleLoginRedeemDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSessionTokensDto'
        '429':
          description: Rate limited
      tags: *ref_6
  /v2/auth/google/{strategy}/exchange:
    get:
      operationId: authGoogleExchangeStartV2
      parameters:
        - name: strategy
          required: true
          in: path
          schema:
            type: string
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicAuthGoogleExchangeDto'
        '429':
          description: Rate limited
      tags: *ref_6
    post:
      operationId: authGoogleExchangeCallbackV2
      parameters:
        - name: strategy
          required: true
          in: path
          schema:
            type: string
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthGoogleExchangeCallbackDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSessionTokensDto'
        '429':
          description: Rate limited
      tags: *ref_6
  /v1/auth/google/{strategy}:
    get:
      operationId: authGoogleExchangeStartV1
      parameters:
        - name: strategy
          required: true
          in: path
          schema:
            type: string
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicAuthGoogleExchangeDto'
        '429':
          description: Rate limited
      tags: *ref_6
    post:
      operationId: authGoogleExchangeCallbackV1
      parameters:
        - name: strategy
          required: true
          in: path
          schema:
            type: string
        - in: header
          name: x-auth-provider
          required: false
          schema:
            type: string
        - in: header
          name: x-auth-strategy
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthGoogleExchangeCallbackDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSessionTokensDto'
        '429':
          description: Rate limited
      tags: *ref_6
...
info:
  title: Fliggs
  description: This is our cool new API
  version: '0.1'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    PublicSessionTokensDto:
      type: object
      properties:
        token:
          type: string
        expires:
          type: string
        refreshToken:
          type: string
        refreshExpires:
          type: string
      required:
        - token
        - expires
        - refreshToken
        - refreshExpires
...
    AuthProvider:
      type: string
      enum:
        - local
        - google
        - apple
        - open-id
    PublicAuthStrategyDto:
      type: object
      properties:
        provider:
          allOf:
            - $ref: '#/components/schemas/AuthProvider'
        name:
          type: string
        title:
          type: string
        tags:
          nullable: true
          type: array
          items:
            type: string
      required:
        - provider
        - name
        - title
...
    AuthAppleLoginRedeemDto:
      type: object
      properties:
        code:
          type: string
      required:
        - code
    PublicAuthAppleExchangeDto:
      type: object
      properties:
        issuerUrl:
          type: string
        clientId:
          type: string
        redirectUrl:
          type: string
        scope:
          type: string
      required:
        - issuerUrl
        - clientId
        - redirectUrl
        - scope
    AuthAppleExchangeCallbackDto:
      type: object
      properties:
        idToken:
          type: string
        firstName:
          type: string
        lastName:
          type: string
      required:
        - idToken
    AuthGoogleLoginRedeemDto:
      type: object
      properties:
        code:
          type: string
      required:
        - code
    PublicAuthGoogleExchangeDto:
      type: object
      properties:
        issuerUrl:
          type: string
        clientId:
          type: string
        redirectUrl:
          type: string
        scope:
          type: string
        webClientId:
          type: string
        iOSClientId:
          type: string
      required:
        - issuerUrl
        - clientId
        - redirectUrl
        - scope
        - webClientId
        - iOSClientId
    AuthGoogleExchangeCallbackDto:
      type: object
      properties:
        idToken:
          type: string
      required:
        - idToken
...